% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casc.R
\name{casc}
\alias{casc}
\title{Covariate Assisted Spectral Clustering}
\usage{
casc(adjMat, covMat, nBlocks, nPoints = 100, method = "regLaplacian",
  rowNorm = F, center = F, verbose = F, assortative = F,
  randStarts = 20)
}
\arguments{
\item{adjMat}{An adjacency matrix}

\item{covMat}{A covariate matrix}

\item{nBlocks}{The number of clusters}

\item{nPoints}{Number of iterations to find the optimal tuning
parameter.}

\item{method}{The form of the adjacency matrix to be used.}

\item{rowNorm}{True if row normalization should be
done before running kmeans.}

\item{center}{A boolean indicating if the covariate matrix columns
should be centered.}

\item{verbose}{A boolean indicating if casc output should include eigendecomposition.}

\item{assortative}{A boolean indicating if the assortative version of casc should be used.}

\item{randStarts}{Number of random restarts for kmeans.}
}
\value{
A list with node cluster assignments, the
the value of the tuning parameter used, the within
cluster sum of squares, and the eigengap.
}
\description{
Covariate Assisted Spectral Clustering
}
\keyword{clustering}
\keyword{spectral}

