% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casc.R
\name{cca}
\alias{cca}
\title{A modified CCA approach for spectral clustering.}
\usage{
cca(adjMat, covMat, nBlocks, method = "regLaplacian", rowNorm = F,
  center = F, verbose = F, randStarts = 10)
}
\arguments{
\item{adjMat}{An adjacency matrix}

\item{covMat}{A covariate matrix}

\item{nBlocks}{The number of clusters}

\item{method}{The form of the adjacency matrix to be used.}

\item{rowNorm}{True if row normalization should be
done before running kmeans.}

\item{center}{A boolean indicating if the covariate matrix columns
should be centered.}

\item{verbose}{A boolean indicating if casc output should include eigendecomposition.}

\item{randStarts}{Number of random restarts for kmeans.}
}
\value{
A list with node cluster assignments, the
the value of the tuning parameter used, the within
cluster sum of squares, and the eigengap.
}
\description{
Uses the eigenvectors of adjMat %*% covMat for spectral clustering.
}
\keyword{clustering}
\keyword{spectral}

