% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misClustRate.R
\name{misClustRate}
\alias{misClustRate}
\title{Calculates the empirical proportion of misclustered nodes.}
\usage{
misClustRate(clusters, nMembers)
}
\arguments{
\item{clusters}{A vector of node cluster memberships}

\item{nMembers}{The number of nodes in each block}
}
\value{
The proportion of misclustered nodes up to identifiability.
}
\description{
Calculates the empirical proportion of misclustered nodes.
}
\keyword{clustering}

