% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partSpecClust}
\alias{partSpecClust}
\title{Performs an approximate spectral clustering on an adjacency matrix by using
the Nystrom extension.}
\usage{
partSpecClust(adjMat, nBlocks, subSampleSize, rowNorm = T, nIter = 10,
  verbose = F)
}
\arguments{
\item{adjMat}{A graph adjacency matrix.}

\item{nBlocks}{The number of clusters in the graph.}

\item{rowNorm}{If true, eigenvector rows should be normalized before
running kmeans (default is true).}

\item{nIter}{Number of kmeans iterations (default is 10).}

\item{verbose}{If true, return a list with cluster assignments, within
cluster sum of squares, and eigendecomposition (default is false).}
}
\value{
A vector of node cluster assignments. Or, if verbose is set to true
a list with cluster assignments and additional information.
}
\description{
Performs an approximate spectral clustering on an adjacency matrix by using
the Nystrom extension.
}
\keyword{clustering}
\keyword{spectral}

