% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotAdj}
\alias{plotAdj}
\title{Plots the adjacency matrix.}
\usage{
plotAdj(adjMat, membership = NULL)
}
\arguments{
\item{adjMat}{The adjacency matrix to be plotted.}

\item{membership}{The node membership used to reorder the adjacency matrix
before plotting.}
}
\description{
Plots the adjacency matrix.
}

