% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCovar.R
\name{simBernCovar}
\alias{simBernCovar}
\title{Simulates a block mixture of Bernoulli covariates.}
\usage{
simBernCovar(covProbMat, nMembers)
}
\arguments{
\item{covProbMat}{A matrix with each row corresponding to a block and
each column corresponding to a covariate. The values are Bernoulli
probabilities.}

\item{nMembers}{The number of nodes in each block.}
}
\value{
A matrix of Bernoulli covariates with a block probability structure.
}
\description{
Simulates a block mixture of Bernoulli covariates.
}
\examples{
covProbMat = matrix(c(.8, .2, .2, .8), nrow = 2)
nMembers = c(50, 50)
simBernCovar(covProbMat, nMembers)
}

