% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simSBM}
\alias{simSBM}
\title{}
\usage{
simSBM(blockPMat, nMembers)
}
\arguments{
\item{blockPMat}{The block probability matrix.}

\item{nMembers}{A vector specifying the number of nodes in each block.}
}
\value{
The sparse representation of an adjacency matrix simulated
according to the Stochastic Blockmodel.
}
\description{

}
\keyword{blockmodel}
\keyword{stochastic}

