% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specClust.R
\name{specClust}
\alias{specClust}
\title{Performs spectral clustering on an adjacency matrix.}
\usage{
specClust(adjMat, nBlocks, method = "regLaplacian", rowNorm = T,
  nIter = 20, verbose = F)
}
\arguments{
\item{adjMat}{A graph adjacency matrix.}

\item{nBlocks}{The number of clusters in the graph.}

\item{method}{The form of the ajacency matrix to be used (default is
regLaplacian).}

\item{rowNorm}{If true, eigenvector rows should be normalized before
running kmeans (default is true).}

\item{nIter}{Number of kmeans iterations (default is 10).}

\item{verbose}{If true, return a list with cluster assignments, within
cluster sum of squares, and eigendecomposition (default is false).}
}
\value{
A vector of node cluster assignments. Or, if verbose is set to true
a list with cluster assignments and additional information.
}
\description{
Performs spectral clustering on an adjacency matrix.
}
\keyword{clustering}
\keyword{spectral}

