% Generated by roxygen2 (4.0.2): do not edit by hand
\name{specClustCov}
\alias{specClustCov}
\title{Spectral clustering of a covariate matrix (low rank).}
\usage{
specClustCov(covMat, nBlocks, nIter = 10)
}
\arguments{
\item{covMat}{A matrix of covariates with each column corresponding to
a covariate.}

\item{nBlocks}{The number of clusters.}

\item{nIter}{Number of kmeans iterations (default is 10).}
}
\value{
Returns a vector of row cluster assignments.
}
\description{
Spectral clustering of a covariate matrix (low rank).
}
\examples{
covProbMat = matrix(c(.6, .2, .2, .6), nrow = 2)
nMembers = c(50, 50)
covMat = simBernCovar(covProbMat, nMembers)
specClustCov(covMat, nBlocks = 2)
}

