% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estSBM}
\alias{estSBM}
\title{Estimates the Stochastic Blockmodel based on the adjacency matrix and given
block membership.}
\usage{
estSBM(adjMat, blockAssign)
}
\arguments{
\item{adjMat}{The adjacency matrix.}

\item{blockAssign}{A vector of node block assignments.}
}
\value{
The estimated block probability matrix from the Stochastic Blockmodel.
}
\description{
Estimates the Stochastic Blockmodel based on the adjacency matrix and given
block membership.
}
\keyword{blockmodel}
\keyword{stochastic}

