% Generated by roxygen2 (4.0.1): do not edit by hand
\name{misClustRate}
\alias{misClustRate}
\title{Calculates the empirical proportion of misclustered nodes.}
\usage{
misClustRate(clusters, nMembers)
}
\arguments{
\item{clusters}{A vector of node cluster memberships}

\item{nMembers}{The number of nodes in each block}
}
\value{
The proportion of misclustered nodes up to identifiablity.
}
\description{
Calculates the empirical proportion of misclustered nodes.
}
\keyword{clustering}

