% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotAdj}
\alias{plotAdj}
\title{Plots the adjacency matrix.}
\usage{
plotAdj(adjMat, membership = NULL, lines = T)
}
\arguments{
\item{adjMat}{The adjacency matrix to be plotted.}

\item{membership}{The node membership used to reorder the adjacency matrix
before plotting.}

\item{lines}{If true, plot lines separating blocks when membership given.}
}
\description{
Plots the adjacency matrix.
}

