% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAdj}
\alias{plotAdj}
\title{Plots the adjacency matrix.}
\usage{
plotAdj(adjMat, membership = NULL)
}
\arguments{
\item{adjMat}{The adjacency matrix to be plotted.}

\item{membership}{The node membership used to reorder the adjacency matrix
before plotting.}
}
\description{
Plots the adjacency matrix.
}

