% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotSBM}
\alias{plotSBM}
\title{Plots the Stochastic Blockmodel.}
\usage{
plotSBM(blockPMat, nMembers = NULL)
}
\arguments{
\item{blockPMat}{The block probability matrix.}

\item{nMembers}{A vector of the number of nodes in each block.}
}
\description{
Plots the Stochastic Blockmodel.
}

