% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSBM}
\alias{plotSBM}
\title{Plots the Stochastic Blockmodel.}
\usage{
plotSBM(blockPMat, nMembers = NULL)
}
\arguments{
\item{blockPMat}{The block probability matrix.}

\item{nMembers}{A vector of the number of nodes in each block.}
}
\description{
Plots the Stochastic Blockmodel.
}

