% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simBernSparseVec}
\alias{simBernSparseVec}
\title{Simulates a Bernoulli random vector.}
\usage{
simBernSparseVec(nElem, p)
}
\arguments{
\item{nElem}{The number of elements in the vector.}

\item{p}{The probability of an element being equal to one.

export}
}
\value{
The sparse representation of a simulated Bernoulli vector.
}
\description{
Simulates a Bernoulli random vector.
}
\keyword{Bernoulli}
\keyword{simulate}
\keyword{sparse}

