% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSBM.R
\name{simSBM}
\alias{simSBM}
\title{Simulates a matrix from the Stochastic Blockmodel.}
\usage{
simSBM(blockPMat, nMembers)
}
\arguments{
\item{blockPMat}{The block probability matrix.}

\item{nMembers}{A vector specifying the number of nodes in each block.}
}
\value{
The sparse representation of an adjacency matrix simulated
according to the Stochastic Blockmodel.
}
\description{
Simulates a matrix from the Stochastic Blockmodel.
}
\keyword{blockmodel}
\keyword{stochastic}

