% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/specClust.R
\name{specClustCov}
\alias{specClustCov}
\title{Spectral clustering of a covariate matrix (low rank).}
\usage{
specClustCov(covMat, nBlocks, nIter = 20, center = F, jit = F)
}
\arguments{
\item{covMat}{A matrix of covariates with each column corresponding to
a covariate.}

\item{nBlocks}{The number of clusters.}

\item{nIter}{Number of kmeans iterations (default is 10).}

\item{center}{A boolean indicating if the covariate matrix columns should
be centered.}
}
\value{
Returns a vector of row cluster assignments.
}
\description{
Spectral clustering of a covariate matrix (low rank).
}
\examples{
covProbMat = matrix(c(.6, .2, .2, .6), nrow = 2)
nMembers = c(50, 50)
covMat = simBernCovar(covProbMat, nMembers)
specClustCov(covMat, nBlocks = 2)
}

